# swagger.yaml
openapi: 3.0.0
info:
  title: raank/xml-reader
  description: 'This is a micro authentication service and users or files crud. '
  contact:
    email: raank92@gmail.com
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8081/api'
paths:
...
  /v1/users:
    get:
      tags:
        - v1.users
      summary: 'All users.'
      operationId: d8f40dfcbbb694dcb7033eebdd31830d
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.paginated.users'
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
      security:
        -
          apiToken: []
    post:
      tags:
        - v1.users
      summary: 'Storing a new User.'
      operationId: aa192f25c50e78c9799d5474566eb83a
      requestBody:
        content:
          application/json:
            schema:
              required:
                - name
                - email
                - password
                - password_confirmation
              properties:
                name:
                  description: 'The name of user.'
                  type: string
                email:
                  description: 'The email of user.'
                  type: string
                password:
                  description: 'The password of user.'
                  type: string
                password_confirmation:
                  description: 'The password confirmation.'
                  type: string
                document:
                  description: 'The document of user.'
                  type: string
              type: object
              example:
                name: 'John Doe'
                email: john@doe.com
                password: password123
                password_confirmation: password123
                document: '12345678'
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                  data: { $ref: '#/components/schemas/v1.models.user' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
      security:
        -
          apiToken: []
  '/v1/users/{userId}':
    get:
      tags:
        - v1.users
      summary: 'Show user specified.'
      operationId: 20b859076cc7f05bb711d52c65c7fa9e
      parameters:
        -
          name: userId
          in: path
          description: 'Identification of User'
          required: true
          schema:
            type: string
          example: 60aeba949828bb0c57abc123
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                  data: { $ref: '#/components/schemas/v1.models.user' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'This information could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notfound'
      security:
        -
          apiToken: []
    put:
      tags:
        - v1.users
      summary: 'Update user specified.'
      operationId: 31dd3427dc5aedb58ce770759e0ccc50
      parameters:
        -
          name: userId
          in: path
          description: 'Identification of User'
          required: true
          schema:
            type: string
          example: 60aeba949828bb0c57abc123
      requestBody:
        content:
          application/json:
            schema:
              properties:
                name:
                  description: 'The name of user.'
                  type: string
                email:
                  description: 'The email of user.'
                  type: string
                document:
                  description: 'The document of user.'
                  type: string
              type: object
              example:
                name: 'John Doe'
                email: john@doe.com
                document: '12345678'
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                  data: { $ref: '#/components/schemas/v1.models.user' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'This information could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notfound'
      security:
        -
          apiToken: []
    delete:
      tags:
        - v1.users
      summary: 'Delete user specified.'
      operationId: 216dd661b9183f9566889fd2e4a1de3a
      parameters:
        -
          name: userId
          in: path
          description: 'Identification of User'
          required: true
          schema:
            type: string
          example: 60aeba949828bb0c57abc123
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                properties:
                  message: { description: 'Message of Response', type: string, example: 'Successful action' }
                type: object
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '404':
          description: 'This information could not be found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notfound'
      security:
        -
          apiToken: []
  /v1/users/search:
    post:
      tags:
        - v1.users
      summary: 'Searching users.'
      operationId: 964381650e5c4e3a91bd8c69b6ba8792
      requestBody:
        content:
          application/json:
            schema:
              properties:
                where:
                  description: 'Where Condition.'
                whereNotNull:
                  description: 'Where field not null.'
                whereNull:
                  description: 'Where field is nullable.'
                orderBy:
                  description: 'The password confirmation.'
                whereBetween:
                  description: 'The where between of user.'
              type: object
              example:
                where:
                  - [field_name, operator, value]
                  - [name, LIKE, john]
                whereNotNull:
                  - field_name
                whereNull:
                  - field_name
                orderBy:
                  - { field_name: field_name, order: DESC }
                whereBetween:
                  - { field_name: [from_value, to_value] }
      responses:
        '200':
          description: 'Successful action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.paginated.users'
        '400':
          description: 'This information could not be processed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: 'You are not authorized for this action'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
      security:
        -
          apiToken: []
...
components:
  schemas:
    BadRequest:
      description: 'This information could not be processed'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'This information could not be processed'
      type: object
    Notfound:
      description: 'This information could not be found'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'This information could not be found'
      type: object
    Unauthorized:
      description: 'You are not authorized for this action'
      properties:
        message:
          description: 'Message of Response'
          type: string
          example: 'You are not authorized for this action'
      type: object
...
    v1.pagination.links:
      description: 'List of Links'
      properties:
        url:
          description: 'URL of Link'
          type: string
          example: 'http://localhost'
        label:
          description: 'Label of Link'
          type: string
          example: my-label
        active:
          description: 'Link is active'
          type: boolean
          example: true
      type: object
...
    v1.paginated.users:
      description: 'Response CRUD paginated'
      allOf:
        -
          $ref: '#/components/schemas/v1.pagination.links'
        -
          properties:
            message:
              description: 'Message of Response'
              type: string
              example: 'Successful action'
            data:
              type: array
              items:
                $ref: '#/components/schemas/v1.models.user'
            current_page:
              description: 'Current page'
              type: integer
              example: 1
            first_page_url:
              description: 'First page URL'
              type: string
              example: 'http://localhost'
            from:
              description: 'From start items'
              type: integer
              example: 1
            last_page:
              description: 'Last Page Number'
              type: integer
              example: 1
            last_page_url:
              description: 'Last page URL'
              type: string
              example: 'http://localhost'
            links:
              description: 'List of Links'
              type: array
              items:
                $ref: '#/components/schemas/v1.pagination.links'
            next_page_url:
              description: 'Next page URL'
              type: string
              example: 'http://localhost'
            path:
              description: 'Path of current URL'
              type: string
              example: 'http://localhost'
            per_page:
              description: 'Items number per page'
              type: integer
              example: 1
            prev_page_url:
              description: 'Prev page URL'
              type: string
              example: 'http://localhost'
            to:
              description: 'Items to end page'
              type: integer
              example: 1
            total:
              description: 'Total of Items'
              type: integer
              example: 1
          type: object
...
    v1.models.user:
      description: 'Response data of user'
      properties:
        id:
          description: 'Identification of User'
          type: integer
          example: 1
        name:
          description: 'Name of User'
          type: string
          example: 'John Doe'
        email:
          description: 'Email of User'
          type: string
          example: john@doe.com
        document:
          description: 'Document of User'
          type: string
          example: '123456789'
        deleted_at:
          description: 'Date of Destroy'
          type: string
          example: null
        updated_at:
          description: 'Date of last updated'
          type: string
          example: '2021-01-01T00:00:00.000000Z'
        created_at:
          description: 'Date of Created'
          type: string
          example: '2021-01-01T00:00:00.000000Z'
      type: object
  securitySchemes:
    apiToken:
      type: apiKey
      name: X-App-Token
      in: header
    bearerAuth:
      type: http
      description: 'Informations of API.'
      scheme: bearer
...